Added russian translation
Added translation for the mod. The translation was verified before uploading, all the main text was translated, except for the dialogue with the authors. Since I don't have all the text, I had to translate it while playing the mod. Therefore, I may have missed part of the text. If you have suggestions to correct the text, write me a message. Thank you for an interesting story mode.